requirements.txt
numpy==1.23.5
uvicorn[standard]
fastapi
python-multipart
face_recognition
pydantic-settings
pytest
pytest-asyncio
httpx
huggingface-hub
